GNG-418: Refactor Result into its own file to simplify Search
